wip: add experimental talvos lifecycle opcodes
The intent is to carve out some opcode space that enables pure-spvasm
compute kernels, i.e. that allows a reasonable "client API" to interpret
the kernel as a stand-alone (homogeneous) program that computes a result
entirely on device.

These experimental opcodes are very much in flux, along with the
(forked) talvos implementation, but they're currently capable of
replacing much of the previous host-side driver DSL in Talvos.

This change introduces:

`OpExecutionGlobalSizeTALVOS`: a directive that expresses to the client
API what the expected dispatch size for a given entry point ought to be,
i.e.

```
OpEntryPoint Kernel %main_fn "main" %gl_GlobalInvocationID

OpExecutionGlobalSizeTALVOS %main_fn 16 1 1
```

Requests that the talvos implementation launch the "main" entrypoint
with a global size of 16x1x1. It's roughly analogous to something like
an `OpExecutionMode GlobalSizeTALVOS ...`.

`OpBufferTALVOS`: a directive to the client API to do a (very limited)
form of I/O allocating an externally accessible, fixed-size buffer with
an optional client API-visible name. e.g.

```
%buf0 = OpBufferTALVOS %_arr_PhysicalStorageBuffer_uint32_t PhysicalStorageBuffer 64 "a"
```

Requests that the client API allocate a 64-byte buffer (named "a") of
the PhysicalStorageBuffer storage class that's associated with the given
type. Allocating it before launch, letting the kernel populate it, and
then doing something useful with it (i.e. copying/dumping it out) before
deallocation after the kernel exits.

It's similar in concept to an `OpVariable`, with a restricted set of
storage classes and types, and an `OpDecorate ... BufferTALVOS 64` plus
an `OpName`.

Both exist to gain experience with implementing and using them, and to
defer decision making around how the needs of the talvos client ought to
fit into the overall spec. They also serve didactic goals as well. By
narrowing the field of potential uses to which a particular SPIR-V
program might be put, they act as an anchor for exploring possible
tradeoffs between their narrow focus and the more general alternatives.